pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
